Make transports able to export promises too
